rename: .page -page로 수정
